// core/types.ts
export enum EducationalStage {
  PRE_SCHOOL = 'PRE_SCHOOL',
  PRIMARY = 'PRIMARY',
  MIDDLE = 'MIDDLE',
  HIGH = 'HIGH',
}

export enum Subject {
  // Primary
  Arabic = 'اللغة العربية',
  French = 'اللغة الفرنسية',
  Math = 'الرياضيات',
  IslamicEducation = 'التربية الإسلامية',
  SocialStudies = 'الاجتماعيات',

  // Middle & High
  English = 'اللغة الإنجليزية',
  Physics = 'الفيزياء والكيمياء',
  NaturalSciences = 'علوم الحياة والأرض',
  ComputerScience = 'المعلوميات',
  Philosophy = 'الفلسفة',
}
// ...
interface BaseContent {
  id: number;
  stage: EducationalStage;
  level: string;
  class: string;
  subject: Subject;
}
// ...
export interface Summary extends BaseContent {
  title: string;
  content: string;
  image?: string;
  pdf?: { name: string; url: string };
  externalLink?: string;
}

export interface Exercise extends BaseContent {
  content: string;
  image?: string;
  pdf?: { name: string; url: string };
  externalLink?: string;
  date: Date;
}